requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
aiohttp==3.9.5
lxml==5.1.0
apscheduler==3.10.4
